+++ x/evm/deps.go
@@ -4,7 +4,6 @@ package evm
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	bank "github.com/cosmos/cosmos-sdk/x/bank/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 )
 
@@ -30,18 +29,6 @@ type AccountKeeper interface {
 	RemoveAccount(ctx sdk.Context, account authtypes.AccountI)
 	GetParams(ctx sdk.Context) (params authtypes.Params)
 	SetModuleAccount(ctx sdk.Context, macc authtypes.ModuleAccountI)
-}
-
-// BankKeeper defines the expected interface needed to retrieve account balances.
-type BankKeeper interface {
-	authtypes.BankKeeper
-	GetBalance(ctx sdk.Context, addr sdk.AccAddress, denom string) sdk.Coin
-	SendCoinsFromModuleToAccount(ctx sdk.Context, senderModule string, recipientAddr sdk.AccAddress, amt sdk.Coins) error
-	MintCoins(ctx sdk.Context, moduleName string, amt sdk.Coins) error
-	BurnCoins(ctx sdk.Context, moduleName string, amt sdk.Coins) error
-
-	GetDenomMetaData(ctx sdk.Context, denom string) (metadata bank.Metadata, isFound bool)
-	SetDenomMetaData(ctx sdk.Context, denomMetaData bank.Metadata)
 }
 
 // StakingKeeper returns the historical headers kept in store.