+++ x/evm/keeper/keeper.go
@@ -15,6 +15,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/codec"
 	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	bankkeeper "github.com/cosmos/cosmos-sdk/x/bank/keeper"
 	gethcommon "github.com/ethereum/go-ethereum/common"
 
 	"github.com/NibiruChain/nibiru/v2/app/appconst"
@@ -40,7 +41,7 @@ type Keeper struct {
 	// this should be the x/gov module account.
 	authority sdk.AccAddress
 
-	bankKeeper    evm.BankKeeper
+	bankKeeper    bankkeeper.Keeper
 	accountKeeper evm.AccountKeeper
 	stakingKeeper evm.StakingKeeper
 
@@ -63,13 +64,14 @@ func NewKeeper(
 	storeKey, transientKey storetypes.StoreKey,
 	authority sdk.AccAddress,
 	accKeeper evm.AccountKeeper,
-	bankKeeper evm.BankKeeper,
+	bankKeeper bankkeeper.Keeper,
 	stakingKeeper evm.StakingKeeper,
 	tracer string,
 ) Keeper {
 	if err := sdk.VerifyAddressFormat(authority); err != nil {
 		panic(err)
 	}
+
 	return Keeper{
 		cdc:           cdc,
 		storeKey:      storeKey,