+++ x/evm/statedb/journal_test.go
@@ -6,10 +6,12 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/MakeNowJust/heredoc/v2"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/core/vm"
 
 	serverconfig "github.com/NibiruChain/nibiru/v2/app/server/config"
+	"github.com/NibiruChain/nibiru/v2/x/common"
 	"github.com/NibiruChain/nibiru/v2/x/common/testutil/testapp"
 	"github.com/NibiruChain/nibiru/v2/x/evm"
 	"github.com/NibiruChain/nibiru/v2/x/evm/embeds"
@@ -18,7 +20,7 @@ import (
 	"github.com/NibiruChain/nibiru/v2/x/evm/statedb"
 )
 
-func (s *Suite) TestPrecompileSnapshots() {
+func (s *Suite) TestComplexJournalChanges() {
 	deps := evmtest.NewTestDeps()
 	bankDenom := evm.EVMBankDenom
 	s.Require().NoError(testapp.FundAccount(
@@ -32,25 +34,11 @@ func (s *Suite) TestPrecompileSnapshots() {
 
 	wasmContract := test.SetupWasmContracts(&deps, &s.Suite)[1]
 	fmt.Printf("wasmContract: %s\n", wasmContract)
-	assertionsBeforeRun := func(deps *evmtest.TestDeps) {
-		test.AssertWasmCounterState(
-			&s.Suite, *deps, wasmContract, 0,
-		)
-	}
-	run := func(deps *evmtest.TestDeps) *vm.EVM {
-		return test.IncrementWasmCounterWithExecuteMulti(
-			&s.Suite, deps, wasmContract, 7,
-		)
-	}
-	assertionsAfterRun := func(deps *evmtest.TestDeps) {
-		test.AssertWasmCounterState(
-			&s.Suite, *deps, wasmContract, 7,
-		)
-	}
 
 	s.T().Log("Assert before transition")
-
-	assertionsBeforeRun(&deps)
+	test.AssertWasmCounterState(
+		&s.Suite, deps, wasmContract, 0,
+	)
 
 	deployArgs := []any{"name", "SYMBOL", uint8(18)}
 	deployResp, err := evmtest.DeployContract(
@@ -71,7 +59,7 @@ func (s *Suite) TestPrecompileSnapshots() {
 
 	s.Run("Populate dirty journal entries. Remove with Commit", func() {
 		stateDB := evmObj.StateDB.(*statedb.StateDB)
-		s.Equal(0, stateDB.DirtiesCount())
+		s.Equal(0, stateDB.DebugDirtiesCount())
 
 		randomAcc := evmtest.NewEthPrivAcc().EthAddr
 		balDelta := evm.NativeToWei(big.NewInt(4))
@@ -81,14 +69,15 @@ func (s *Suite) TestPrecompileSnapshots() {
 		stateDB.AddBalance(randomAcc, balDelta)
 		// 1 dirties from [balanceChange]
 		stateDB.SubBalance(randomAcc, balDelta)
-		if stateDB.DirtiesCount() != 4 {
+		if stateDB.DebugDirtiesCount() != 4 {
 			debugDirtiesCountMismatch(stateDB, s.T())
 			s.FailNow("expected 4 dirty journal changes")
 		}
 
-		err = stateDB.Commit() // Dirties should be gone
+		s.T().Log("StateDB.Commit, then Dirties should be gone")
+		err = stateDB.Commit()
 		s.NoError(err)
-		if stateDB.DirtiesCount() != 0 {
+		if stateDB.DebugDirtiesCount() != 0 {
 			debugDirtiesCountMismatch(stateDB, s.T())
 			s.FailNow("expected 0 dirty journal changes")
 		}
@@ -110,7 +99,7 @@ func (s *Suite) TestPrecompileSnapshots() {
 		)
 		s.Require().NoError(err)
 		stateDB := evmObj.StateDB.(*statedb.StateDB)
-		if stateDB.DirtiesCount() != 2 {
+		if stateDB.DebugDirtiesCount() != 2 {
 			debugDirtiesCountMismatch(stateDB, s.T())
 			s.FailNow("expected 2 dirty journal changes")
 		}
@@ -136,22 +125,105 @@ func (s *Suite) TestPrecompileSnapshots() {
 		s.Require().ErrorContains(err, vm.ErrExecutionReverted.Error())
 	})
 
-	s.Run("Precompile calls also start and end clean (no dirty changes)", func() {
-		evmObj = run(&deps)
-		assertionsAfterRun(&deps)
+	s.Run("Precompile calls populate snapshots", func() {
+		s.T().Log("commitEvmTx=true, expect 0 dirty journal entries")
+		commitEvmTx := true
+		evmObj = test.IncrementWasmCounterWithExecuteMulti(
+			&s.Suite, &deps, wasmContract, 7, commitEvmTx,
+		)
+		// assertions after run
+		test.AssertWasmCounterState(
+			&s.Suite, deps, wasmContract, 7,
+		)
 		stateDB, ok := evmObj.StateDB.(*statedb.StateDB)
 		s.Require().True(ok, "error retrieving StateDB from the EVM")
-		if stateDB.DirtiesCount() != 0 {
+		if stateDB.DebugDirtiesCount() != 0 {
 			debugDirtiesCountMismatch(stateDB, s.T())
 			s.FailNow("expected 0 dirty journal changes")
 		}
+
+		s.T().Log("commitEvmTx=false, expect dirty journal entries")
+		commitEvmTx = false
+		evmObj = test.IncrementWasmCounterWithExecuteMulti(
+			&s.Suite, &deps, wasmContract, 5, commitEvmTx,
+		)
+		stateDB, ok = evmObj.StateDB.(*statedb.StateDB)
+		s.Require().True(ok, "error retrieving StateDB from the EVM")
+
+		s.T().Log("Expect exactly 0 dirty journal entry for the precompile snapshot")
+		if stateDB.DebugDirtiesCount() != 0 {
+			debugDirtiesCountMismatch(stateDB, s.T())
+			s.FailNow("expected 0 dirty journal changes")
+		}
+
+		s.T().Log("Expect no change since the StateDB has not been committed")
+		test.AssertWasmCounterState(
+			&s.Suite, deps, wasmContract, 7, // 7 = 7 + 0
+		)
+
+		s.T().Log("Expect change to persist on the StateDB cacheCtx")
+		cacheCtx := stateDB.GetCacheContext()
+		s.NotNil(cacheCtx)
+		deps.Ctx = *cacheCtx
+		test.AssertWasmCounterState(
+			&s.Suite, deps, wasmContract, 12, // 12 = 7 + 5
+		)
+		// NOTE: that the [StateDB.Commit] fn has not been called yet. We're still
+		// mid-transaction.
+
+		s.T().Log("EVM revert operation should bring about the old state")
+		err = test.IncrementWasmCounterWithExecuteMultiViaVMCall(
+			&s.Suite, &deps, wasmContract, 50, commitEvmTx, evmObj,
+		)
+		stateDBPtr := evmObj.StateDB.(*statedb.StateDB)
+		s.Require().Equal(stateDB, stateDBPtr)
+		s.Require().NoError(err)
+		s.T().Log(heredoc.Doc(`At this point, 2 precompile calls have succeeded.
+One that increments the counter to 7 + 5, and another for +50. 
+The StateDB has not been committed. We expect to be able to revert to both
+snapshots and see the prior states.`))
+		cacheCtx = stateDB.GetCacheContext()
+		deps.Ctx = *cacheCtx
+		test.AssertWasmCounterState(
+			&s.Suite, deps, wasmContract, 7+5+50,
+		)
+
+		errFn := common.TryCatch(func() {
+			// There were only two EVM calls.
+			// Thus, there are only 2 snapshots: 0 and 1.
+			// We should not be able to revert to a third one.
+			stateDB.RevertToSnapshot(2)
+		})
+		s.Require().ErrorContains(errFn(), "revision id 2 cannot be reverted")
+
+		stateDB.RevertToSnapshot(1)
+		cacheCtx = stateDB.GetCacheContext()
+		s.NotNil(cacheCtx)
+		deps.Ctx = *cacheCtx
+		test.AssertWasmCounterState(
+			&s.Suite, deps, wasmContract, 7+5,
+		)
+
+		stateDB.RevertToSnapshot(0)
+		cacheCtx = stateDB.GetCacheContext()
+		s.NotNil(cacheCtx)
+		deps.Ctx = *cacheCtx
+		test.AssertWasmCounterState(
+			&s.Suite, deps, wasmContract, 7, // state before precompile called
+		)
+
+		err = stateDB.Commit()
+		deps.Ctx = stateDB.GetEvmTxContext()
+		test.AssertWasmCounterState(
+			&s.Suite, deps, wasmContract, 7, // state before precompile called
+		)
 	})
 }
 
 func debugDirtiesCountMismatch(db *statedb.StateDB, t *testing.T) string {
 	lines := []string{}
-	dirties := db.Dirties()
-	stateObjects := db.StateObjects()
+	dirties := db.DebugDirties()
+	stateObjects := db.DebugStateObjects()
 	for addr, dirtyCountForAddr := range dirties {
 		lines = append(lines, fmt.Sprintf("Dirty addr: %s, dirtyCountForAddr=%d", addr, dirtyCountForAddr))